added api of screening

+++ server/src/main/java/com/recruitment/server/repository/UserRepository.java
@@ -1,12 +1,12 @@
 package com.recruitment.server.repository;
 
-
 import com.recruitment.server.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
+
     boolean existsByEmail(String email);
 }
